requirements.txt
Django==5.2.7
django-environ==0.12.0
django-cors-headers==4.9.0
graphene-django==3.2.3
django-tinymce==4.1.0
django-modeltranslation==1.19.2